services/telemetry-ingestion/requirements.txt
kafka-python==2.0.2
clickhouse-driver==0.2.6
jsonschema==4.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-json-logger==2.0.7
